fix use after move in websocket handshake method
This fixes an issue in which the socket would never return the response
string during the handshake phase of a websocket connection
initialization.

This was due to the fact that the response_str variable was moved into
the capture scope of the WriteHandler lambda passed as the third
argument to async_write.

Since this variable isn' used within the context of the lambda, there
doesn't seem to be need for it anyhow.