fix use after move in websocket handshake method (#1857)
* fix use after move in websocket handshake method

This fixes an issue in which the socket would never return the response
string during the handshake phase of a websocket connection
initialization.

This was due to the fact that the response_str variable was moved into
the capture scope of the WriteHandler lambda passed as the third
argument to async_write.

Since this variable isn' used within the context of the lambda, there
doesn't seem to be need for it anyhow.

* Use shared_ptr to keep memory alive

* fix name

---------

Co-authored-by: Stephen Berry <stephenberry.developer@gmail.com>