Add title and required to generated JSON schema (#1850)
* FEATURE: [schema] Mark non-nullable members as required

Currently, when generating code based on the generated schema files, all
members become optional with typical JSON schema code generators and
languages like C++ or rust as output target. To better reflect the
semantics of the code, this change marks non-nullable object members as
required. To do so, enable the option `error_on_missing_keys` option.

* FEATURE: [schema] Set the name of the type as title

Some JSON schema code generators like oxidecomputer/typify use the title
of a schema as the name of the generated type. Thus, this change sets a
sane default for the title.

* FEATURE: [schema] Add requires_key functionallity

Allow per-type customization if a member is required or not via
meta<T>::requires_key(key, is_nullable).

* spelling fix and adding constexpr to requires_key call

* Fix compiler bug

* Update json-schema.md

---------

Co-authored-by: Olaf Leidinger <oleid@mescharet.de>
Co-authored-by: Stephen Berry <stephenberry.developer@gmail.com>